Clean up imports and naming
refs #30819
